move nanomina_to_mina to utility module
