+++ rust/src/utility/mod.rs
@@ -1,0 +1,1 @@
+pub mod functions;
